rename: 패키지 변경, Request, Response 역할 추가
