--- src/main/java/com/cheocharm/MapZ/diary/domain/dto/DiaryLikePeopleDto.java
@@ -1,11 +1,0 @@
-package com.cheocharm.MapZ.diary.domain.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class DiaryLikePeopleDto {
-    private String userImageUrl;
-    private String username;
-}
