--- src/main/java/com/cheocharm/MapZ/diary/domain/dto/LikeDiaryDto.java
@@ -1,12 +1,0 @@
-package com.cheocharm.MapZ.diary.domain.dto;
-
-import lombok.Getter;
-
-import javax.validation.constraints.NotNull;
-
-@Getter
-public class LikeDiaryDto {
-
-    @NotNull
-    private Long diaryId;
-}
