--- src/main/java/com/cheocharm/MapZ/diary/domain/dto/request/DeleteTempDiaryRequest.java
@@ -1,11 +1,0 @@
-package com.cheocharm.MapZ.diary.domain.dto.request;
-
-import lombok.Getter;
-
-import javax.validation.constraints.NotNull;
-
-@Getter
-public class DeleteTempDiaryRequest {
-    @NotNull
-    private Long diaryId;
-}
