--- src/main/java/com/cheocharm/MapZ/diary/domain/dto/request/WriteDiaryImageRequest.java
@@ -1,20 +1,0 @@
-package com.cheocharm.MapZ.diary.domain.dto.request;
-
-import lombok.Getter;
-
-import javax.validation.constraints.NotNull;
-
-@Getter
-public class WriteDiaryImageRequest {
-    @NotNull
-    private Long groupId;
-
-    @NotNull
-    private String address;
-
-    @NotNull
-    private Double latitude;
-
-    @NotNull
-    private Double longitude;
-}
