+++ src/main/java/com/cheocharm/MapZ/diary/presentation/dto/request/DeleteDiaryRequest.java
@@ -1,0 +1,14 @@
+package com.cheocharm.MapZ.diary.presentation.dto.request;
+
+import lombok.Getter;
+
+import javax.validation.constraints.NotNull;
+
+@Getter
+public class DeleteDiaryRequest {
+    @NotNull
+    private Long userId;
+
+    @NotNull
+    private Long diaryId;
+}
