+++ src/main/java/com/cheocharm/MapZ/diary/presentation/dto/request/DeleteTempDiaryRequest.java
@@ -1,0 +1,11 @@
+package com.cheocharm.MapZ.diary.presentation.dto.request;
+
+import lombok.Getter;
+
+import javax.validation.constraints.NotNull;
+
+@Getter
+public class DeleteTempDiaryRequest {
+    @NotNull
+    private Long diaryId;
+}
