+++ src/main/java/com/cheocharm/MapZ/diary/presentation/dto/request/WriteDiaryRequest.java
@@ -1,0 +1,18 @@
+package com.cheocharm.MapZ.diary.presentation.dto.request;
+
+import lombok.Getter;
+
+import javax.validation.constraints.NotNull;
+
+@Getter
+public class WriteDiaryRequest {
+
+    @NotNull
+    private String title;
+
+    @NotNull
+    private String content;
+
+    @NotNull
+    private Long diaryId;
+}
