--- src/main/java/com/cheocharm/MapZ/group/domain/dto/ChangeChiefDto.java
@@ -1,15 +1,0 @@
-package com.cheocharm.MapZ.group.domain.dto;
-
-import lombok.Getter;
-
-import javax.validation.constraints.NotNull;
-
-@Getter
-public class ChangeChiefDto {
-
-    @NotNull
-    private Long groupId;
-
-    @NotNull
-    private Long userId;
-}
