--- src/main/java/com/cheocharm/MapZ/group/domain/dto/ChangeGroupInfoDto.java
@@ -1,21 +1,0 @@
-package com.cheocharm.MapZ.group.domain.dto;
-
-import lombok.Getter;
-
-import javax.validation.constraints.NotNull;
-
-@Getter
-public class ChangeGroupInfoDto {
-    @NotNull
-    private Long groupId;
-
-    @NotNull
-    private String groupName;
-
-    @NotNull
-    private String bio;
-
-    @NotNull
-    private Boolean changeStatus;
-
-}
