--- src/main/java/com/cheocharm/MapZ/group/domain/dto/ExitGroupDto.java
@@ -1,12 +1,0 @@
-package com.cheocharm.MapZ.group.domain.dto;
-
-import lombok.Getter;
-
-import javax.validation.constraints.NotNull;
-
-@Getter
-public class ExitGroupDto {
-
-    @NotNull
-    Long groupId;
-}
