--- src/main/java/com/cheocharm/MapZ/group/domain/dto/GetGroupListDto.java
@@ -1,15 +1,0 @@
-package com.cheocharm.MapZ.group.domain.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class GetGroupListDto {
-
-    private String groupName;
-    private String groupImageUrl;
-    private Long count;
-    private String chiefUserImage;
-    private Long groupId;
-}
