--- src/main/java/com/cheocharm/MapZ/group/domain/dto/GroupMemberDto.java
@@ -1,14 +1,0 @@
-package com.cheocharm.MapZ.group.domain.dto;
-
-import com.cheocharm.MapZ.usergroup.InvitationStatus;
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-public class GroupMemberDto {
-    private String username;
-    private String userImageUrl;
-    private Long userId;
-    private InvitationStatus invitationStatus;
-}
