--- src/main/java/com/cheocharm/MapZ/group/domain/dto/InviteUserListDto.java
@@ -1,16 +1,0 @@
-package com.cheocharm.MapZ.group.domain.dto;
-
-import lombok.Getter;
-
-import javax.validation.constraints.NotNull;
-import java.util.List;
-
-@Getter
-public class InviteUserListDto {
-
-    @NotNull
-    private Long groupId;
-
-    @NotNull
-    private List<Long> userIdList;
-}
