--- src/main/java/com/cheocharm/MapZ/group/domain/dto/JoinGroupDto.java
@@ -1,13 +1,0 @@
-package com.cheocharm.MapZ.group.domain.dto;
-
-import lombok.Getter;
-
-import javax.validation.constraints.NotNull;
-
-@Getter
-public class JoinGroupDto {
-
-    @NotNull
-    private Long groupId;
-
-}
