--- src/main/java/com/cheocharm/MapZ/group/domain/dto/JoinGroupResultDto.java
@@ -1,11 +1,0 @@
-package com.cheocharm.MapZ.group.domain.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class JoinGroupResultDto {
-    private Boolean alreadyJoin;
-    private String status;
-}
