+++ src/main/java/com/cheocharm/MapZ/group/presentation/dto/request/ExitGroupRequest.java
@@ -1,0 +1,12 @@
+package com.cheocharm.MapZ.group.presentation.dto.request;
+
+import lombok.Getter;
+
+import javax.validation.constraints.NotNull;
+
+@Getter
+public class ExitGroupRequest {
+
+    @NotNull
+    Long groupId;
+}
