+++ src/main/java/com/cheocharm/MapZ/group/presentation/dto/request/KickUserRequest.java
@@ -1,0 +1,14 @@
+package com.cheocharm.MapZ.group.presentation.dto.request;
+
+import lombok.Getter;
+
+import javax.validation.constraints.NotNull;
+
+@Getter
+public class KickUserRequest {
+    @NotNull
+    private Long userId;
+
+    @NotNull
+    private Long groupId;
+}
