+++ src/main/java/com/cheocharm/MapZ/group/presentation/dto/response/GroupMemberResponse.java
@@ -1,0 +1,14 @@
+package com.cheocharm.MapZ.group.presentation.dto.response;
+
+import com.cheocharm.MapZ.usergroup.domain.InvitationStatus;
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class GroupMemberResponse {
+    private String username;
+    private String userImageUrl;
+    private Long userId;
+    private InvitationStatus invitationStatus;
+}
