+++ src/main/java/com/cheocharm/MapZ/like/presentation/dto/request/LikeDiaryRequest.java
@@ -1,0 +1,12 @@
+package com.cheocharm.MapZ.like.presentation.dto.request;
+
+import lombok.Getter;
+
+import javax.validation.constraints.NotNull;
+
+@Getter
+public class LikeDiaryRequest {
+
+    @NotNull
+    private Long diaryId;
+}
