+++ src/main/java/com/cheocharm/MapZ/like/presentation/dto/response/DiaryLikePeopleResponse.java
@@ -1,0 +1,11 @@
+package com.cheocharm.MapZ.like.presentation.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class DiaryLikePeopleResponse {
+    private String userImageUrl;
+    private String username;
+}
