--- src/main/java/com/cheocharm/MapZ/user/domain/dto/GoogleIdTokenDto.java
@@ -1,20 +1,0 @@
-package com.cheocharm.MapZ.user.domain.dto;
-
-import lombok.Getter;
-
-@Getter
-public class GoogleIdTokenDto {
-    private String iss;
-    private String sub;
-    private String azp;
-    private String aud;
-    private String iat;
-    private String exp;
-    private String email;
-    private String email_verified;
-    private String name;
-    private String picture;
-    private String given_name;
-    private String family_name;
-    private String locale;
-}
