--- src/main/java/com/cheocharm/MapZ/user/domain/dto/GoogleSignUpDto.java
@@ -1,18 +1,0 @@
-package com.cheocharm.MapZ.user.domain.dto;
-
-import lombok.Getter;
-
-import javax.validation.constraints.NotNull;
-
-@Getter
-public class GoogleSignUpDto {
-
-    @NotNull
-    private String username;
-
-    @NotNull
-    private String idToken;
-
-    @NotNull
-    private Boolean pushAgreement;
-}
