--- src/main/java/com/cheocharm/MapZ/user/domain/dto/MapZSignInDto.java
@@ -1,15 +1,0 @@
-package com.cheocharm.MapZ.user.domain.dto;
-
-import lombok.Getter;
-
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotNull;
-
-@Getter
-public class MapZSignInDto {
-    @NotNull @Email
-    private String email;
-
-    @NotNull
-    private String password;
-}
