--- src/main/java/com/cheocharm/MapZ/user/domain/dto/MyPageInfoDto.java
@@ -1,11 +1,0 @@
-package com.cheocharm.MapZ.user.domain.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class MyPageInfoDto {
-    private String username;
-    private String userImageUrl;
-}
