--- src/main/java/com/cheocharm/MapZ/user/domain/dto/TokenPairResponseDto.java
@@ -1,11 +1,0 @@
-package com.cheocharm.MapZ.user.domain.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class TokenPairResponseDto {
-    private String accessToken;
-    private String refreshToken;
-}
