+++ src/main/java/com/cheocharm/MapZ/user/presentation/dto/response/GetUserListResponse.java
@@ -1,0 +1,26 @@
+package com.cheocharm.MapZ.user.presentation.dto.response;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+import java.util.List;
+
+@Getter
+@AllArgsConstructor
+public class GetUserListResponse {
+    private Boolean hasNext;
+    private List<UserList> userList;
+
+    @Getter
+    @Builder
+    public static class UserList {
+        private String username;
+        private String userImageUrl;
+        private Long userId;
+        @Schema(description = "그룹에 포함되어 있는 멤버면 true, 아니면 false")
+        private Boolean member;
+    }
+
+}
