+++ src/main/java/com/cheocharm/MapZ/user/presentation/dto/response/MyPageInfoResponse.java
@@ -1,0 +1,11 @@
+package com.cheocharm.MapZ.user.presentation.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class MyPageInfoResponse {
+    private String username;
+    private String userImageUrl;
+}
