+++ src/main/java/com/cheocharm/MapZ/user/presentation/dto/response/TokenPairResponse.java
@@ -1,0 +1,11 @@
+package com.cheocharm.MapZ.user.presentation.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class TokenPairResponse {
+    private String accessToken;
+    private String refreshToken;
+}
